requirements.txt
urllib3==1.26.9
requests==2.27.1
python-telegram-bot==13.11